+++ src/types/hcl/hcl_data.go
@@ -1,0 +1,4 @@
+package hcl
+
+// Data represents a terraform data block in a File
+type Data map[string]interface{}